fix(regionenkarte): make regionenkarte legend reusable as its own component
