+++ src/lang/de.json
@@ -531,6 +531,10 @@
   "Rauszoomen": "Rauszoomen",
   "Rechtliches": "Rechtliches",
   "Region": "Region",
+  "Region Mitte": "Region Mitte",
+  "Region Ost": "Region Ost",
+  "Region Süd": "Region Süd",
+  "Region West": "Region West",
   "Regionaler Naturpark": "Regionaler Naturpark",
   "Richtung": "Richtung",
   "rollstuhl_billet": "Billettschalter mit Rollstuhl zugänglich",
