+++ src/lang/en.json
@@ -568,6 +568,10 @@
   "Rauszoomen": "Zoom out",
   "Rechtliches": "Legal information",
   "Region": "Region",
+  "Region Mitte": "Central region",
+  "Region Ost": "Eastern region",
+  "Region Süd": "Southern region",
+  "Region West": "Western region",
   "Regionaler Naturpark": "Regional nature park",
   "Richtung": "Heading",
   "rollstuhl_billet": "Wheelchair ticket",
