+++ src/lang/fr.json
@@ -603,6 +603,10 @@
   "Rauszoomen": "Réduire",
   "Rechtliches": "Mention juridique",
   "Region": "Région",
+  "Region Mitte": "Région Centre",
+  "Region Ost": "Région Est",
+  "Region Süd": "Région Sud",
+  "Region West": "Région Ouest",
   "Regionaler Naturpark": "Parc naturel régional",
   "Richtung": "Direction",
   "Rohfilter": "Filtre brut",
