+++ src/lang/it.json
@@ -605,6 +605,10 @@
   "Rauszoomen": "Rimpiccolire",
   "Rechtliches": "Note legali",
   "Region": "Regione",
+  "Region Mitte": "Regione centrale",
+  "Region Ost": "Regione est",
+  "Region Süd": "Regione sud",
+  "Region West": "Regione ovest",
   "Regionaler Naturpark": "Parco naturale regionale",
   "Richtung": "Direzione",
   "Rohfilter": "Filtro generico",
